+++ src/components/TenantView/index.ts
@@ -1,0 +1,1 @@
+export { TenantView } from './TenantView';
